+++ libraries/KT0803/src/KT0803.h
@@ -2,8 +2,8 @@
 //
 //    FILE: KT0803.h
 //  AUTHOR: Rob Tillaart
-// VERSION: 0.1.0
-// PURPOSE: Arduino Library for KT0803 FM transmitter
+// VERSION: 0.2.0
+// PURPOSE: Arduino Library for KT0803 and KT0803K FM transmitter
 //     URL: https://github.com/RobTillaart/KT0803
 
 
@@ -11,7 +11,7 @@
 #include "Wire.h"
 
 
-#define KT0803_LIB_VERSION          (F("0.1.0"))
+#define KT0803_LIB_VERSION          (F("0.2.0"))
 
 
 class KT0803
@@ -19,14 +19,14 @@ class KT0803
 public:
   KT0803(TwoWire * wire = &Wire);
 
-  bool begin();
+  bool begin(float freq = 90.0, bool mute = true);
   bool isConnected();
 
 
   //  FM FREQUENCY
-  bool setFrequency(float frequency);
-  float getFrequency();
-  bool setChannel(uint16_t channel);
+  bool     setFrequency(float MHz);
+  float    getFrequency();
+  bool     setChannel(uint16_t channel);
   uint16_t getChannel();
 
 
@@ -42,7 +42,7 @@ public:
   // 011: -12dB
   uint8_t getPGA();
 
-  bool setRFGain(uint8_t rfgain);
+  bool setRFGain(uint8_t rfgain);  //  0-15
   uint8_t getRFgain();
 
 
@@ -63,7 +63,7 @@ public:
 
   //  MUTE software
   bool setMute(bool mute);  //  true == muted
-  bool getMute();         //  isMuted().
+  bool getMute();           //  isMuted().
 
 
 protected:
@@ -73,6 +73,21 @@ protected:
  
   uint8_t   _address = 0x3E;  //  fixed address for KT0803.
   TwoWire * _wire = NULL;
+};
+
+
+/////////////////////////////////////////////////////////////////////////////
+//
+//  DERIVED CLASSES
+//
+class KT0803K : public KT0803
+{
+public:
+  KT0803K(TwoWire * wire = &Wire);
+
+  //  CHANNEL
+  bool     setChannel(uint16_t channel);
+  uint16_t getChannel();
 };