+++ libraries/KT0803/CHANGELOG.md
@@ -6,5 +6,18 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
 
+## [0.2.0] - 2024-03-09
+- Fix #2 frequency to channel formula
+- add derived **class KT0803K**
+- add examples
+- add parameter checks functions.
+- add default frequency + mute to **begin()**
+- update documentation
+- update GitHub actions
+- minor edits
+
+
+----
+
 ## [0.1.0] - 2023-12-27
 - initial version